# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deriv Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deriv/connect:
    post:
      operationId: connectDeriv
      tags: [deriv]
      summary: Save Deriv API token for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DerivTokenInput"
      responses:
        "200":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DerivStatus"

    delete:
      operationId: disconnectDeriv
      tags: [deriv]
      summary: Remove Deriv API token
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DerivStatus"
...
  /deriv/active-symbols:
    get:
      operationId: searchDerivSymbols
      tags: [deriv]
      summary: Search Deriv active symbols (all tradeable instruments)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: instrumentType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching symbols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DerivActiveSymbol"

  # ─── Assets ──────────────────────────────────────────────────────────────────
...
  /autotrade/sessions:
    get:
      operationId: listAutoTradeSessions
      tags: [autotrade]
      summary: List auto trading sessions for the current user
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoTradeSession"

    post:
      operationId: createAutoTradeSession
      tags: [autotrade]
      summary: Start a new auto trading session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoTradeSessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoTradeSession"
  /autotrade/sessions/{id}:
    patch:
      operationId: updateAutoTradeSession
      tags: [autotrade]
      summary: Update auto trading session status (pause/stop/resume)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoTradeSessionUpdate"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoTradeSession"

    delete:
      operationId: deleteAutoTradeSession
      tags: [autotrade]
      summary: Delete auto trading session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DerivTokenInput:
      type: object
      properties:
        apiToken:
          type: string
        appId:
          type: ["string", "null"]
        accountId:
          type: ["string", "null"]
      required:
        - apiToken
    DerivStatus:
      type: object
      properties:
        connected:
          type: boolean
        appId:
          type: ["string", "null"]
        accountId:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        balance:
          type: ["number", "null"]
        loginId:
          type: ["string", "null"]
        connectedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - connected

    # ─── Assets ─────────────────────────────────────────────────────────────────
...
    AutoTradeSession:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        strategyId:
          type: integer
        strategyName:
          type: ["string", "null"]
        status:
          type: string
          enum: [running, stopped, paused, error]
        mode:
          type: string
          enum: [demo, live]
        symbol:
          type: string
        stakeAmount:
          type: number
        duration:
          type: integer
        durationUnit:
          type: string
        maxTrades:
          type: ["integer", "null"]
        stopOnLoss:
          type: ["number", "null"]
        symbols:
          type: array
          items:
            type: string
        pairMode:
          type: string
          enum: [single, rotating, simultaneous]
        profitTarget:
          type: ["number", "null"]
        totalTrades:
          type: integer
        winTrades:
          type: integer
        totalPnl:
          type: number
        errorMessage:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        stoppedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - strategyId
        - status
        - mode
        - symbol
        - stakeAmount
        - duration
        - durationUnit
        - totalTrades
        - winTrades
        - totalPnl
        - startedAt
        - createdAt
    AutoTradeSessionInput:
      type: object
      properties:
        strategyId:
          type: integer
        mode:
          type: string
          enum: [demo, live]
        symbol:
          type: string
        symbols:
          type: array
          items:
            type: string
        pairMode:
          type: string
          enum: [single, rotating, simultaneous]
        stakeAmount:
          type: number
        duration:
          type: integer
        durationUnit:
          type: string
        maxTrades:
          type: ["integer", "null"]
        stopOnLoss:
          type: ["number", "null"]
        profitTarget:
          type: ["number", "null"]
      required:
        - strategyId
        - mode
        - symbol
        - stakeAmount
        - duration
        - durationUnit
    AutoTradeSessionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [running, stopped, paused]
        stakeAmount:
          type: number
        duration:
          type: integer
        durationUnit:
          type: string
        symbols:
          type: array
          items:
            type: string
        pairMode:
          type: string
          enum: [single, rotating, simultaneous]
        maxTrades:
          type: ["integer", "null"]
        stopOnLoss:
          type: ["number", "null"]
        profitTarget:
          type: ["number", "null"]

    # ─── Deriv Symbols ───────────────────────────────────────────────────────────
    DerivActiveSymbol:
      type: object
      properties:
        symbol:
          type: string
        displayName:
          type: string
        shortName:
          type: string
        instrumentType:
          type: string
        subtype:
          type: ["string", "null"]
        isTradingSuspended:
          type: boolean
        pip:
          type: ["number", "null"]
      required:
        - symbol
        - displayName
        - shortName
        - instrumentType
        - isTradingSuspended
...