# lib/api-zod/src/generated/types/autoTradeSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AutoTradeSessionMode } from './autoTradeSessionMode';
import type { AutoTradeSessionPairMode } from './autoTradeSessionPairMode';
import type { AutoTradeSessionStatus } from './autoTradeSessionStatus';

export interface AutoTradeSession {
  id: number;
  userId: string;
  strategyId: number;
  /** @nullable */
  strategyName?: string | null;
  status: AutoTradeSessionStatus;
  mode: AutoTradeSessionMode;
  symbol: string;
  stakeAmount: number;
  duration: number;
  durationUnit: string;
  /** @nullable */
  maxTrades?: number | null;
  /** @nullable */
  stopOnLoss?: number | null;
  symbols?: string[];
  pairMode?: AutoTradeSessionPairMode;
  /** @nullable */
  profitTarget?: number | null;
  totalTrades: number;
  winTrades: number;
  totalPnl: number;
  /** @nullable */
  errorMessage?: string | null;
  startedAt: Date;
  /** @nullable */
  stoppedAt?: Date | null;
  createdAt: Date;
}