# lib/api-zod/src/generated/types/autoTradeSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AutoTradeSessionInputMode } from './autoTradeSessionInputMode';
import type { AutoTradeSessionInputPairMode } from './autoTradeSessionInputPairMode';

export interface AutoTradeSessionInput {
  strategyId: number;
  mode: AutoTradeSessionInputMode;
  symbol: string;
  symbols?: string[];
  pairMode?: AutoTradeSessionInputPairMode;
  stakeAmount: number;
  duration: number;
  durationUnit: string;
  /** @nullable */
  maxTrades?: number | null;
  /** @nullable */
  stopOnLoss?: number | null;
  /** @nullable */
  profitTarget?: number | null;
}