# lib/api-zod/src/generated/types/derivStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deriv Trading Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DerivStatus {
  connected: boolean;
  /** @nullable */
  appId?: string | null;
  /** @nullable */
  accountId?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  balance?: number | null;
  /** @nullable */
  loginId?: string | null;
  /** @nullable */
  connectedAt?: Date | null;
}